+++ bin/ci-scripts/set_env_with_etcd.sh
@@ -34,7 +34,7 @@ then
 fi
 
 if [ -z "${BUILDENV}" ]
-then 
+then
     BUILDENV="dev"
 fi
 
@@ -43,6 +43,26 @@ then
     SYMFONYENV=${BUILDENV}
 fi
 
+if [ -z "$DatabaseName" ]
+then
+    DatabaseName=pia_db_$Suffix
+fi
+
+if [ -z "$DatabaseUser" ]
+then
+    DatabaseUser=pia_user_$Suffix
+fi
+
+if [ -z "$DatabasePassword" ]
+then
+    DatabasePassword=pia_user_$Suffix
+fi
+
+if [ -z "$MailerUrl" ]
+then
+    MailerUrl=smtp://127.0.0.1:1025
+
+fi
 
 
 
@@ -61,12 +81,14 @@ $ETCDCTLCMD put $Prefix/postgres/hostname $postgreshost $ETCDENDPOINT
 $ETCDCTLCMD put $Prefix/postgres/root/username $postgresuser $ETCDENDPOINT
 $ETCDCTLCMD put $Prefix/postgres/root/password $postgrespass $ETCDENDPOINT
 
-$ETCDCTLCMD put $Prefix/postgres/default/dbname pia_db_$Suffix $ETCDENDPOINT
-$ETCDCTLCMD put $Prefix/postgres/default/username pia_user_$Suffix $ETCDENDPOINT
-$ETCDCTLCMD put $Prefix/postgres/default/password pia_pass_$Suffix $ETCDENDPOINT
+$ETCDCTLCMD put $Prefix/postgres/default/dbname $DatabaseName $ETCDENDPOINT
+$ETCDCTLCMD put $Prefix/postgres/default/username $DatabaseUser $ETCDENDPOINT
+$ETCDCTLCMD put $Prefix/postgres/default/password $DatabasePassword $ETCDENDPOINT
+
+$ETCDCTLCMD put $Prefix/smtp/default/url $MailerUrl $ETCDENDPOINT
 
 # set symfony env
-$ETCDCTLCMD put $Prefix/symfony/env $SYMFONYENV $ETCDENDPOINT 
+$ETCDCTLCMD put $Prefix/symfony/env $SYMFONYENV $ETCDENDPOINT
 # get ip
 currentip=$(hostname -i) # works only if the host name can be resolved
 $ETCDCTLCMD put $Prefix/url/addr $currentip':8000' $ETCDENDPOINT