+++ src/Auth/ClientManager.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * Copyright (C) 2015-2018 Libre Informatique
+ *
+ * This file is licenced under the GNU LGPL v3.
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
+namespace PiaApi\Auth;
+
+use FOS\OAuthServerBundle\Entity\ClientManager as BaseClientManager;
+
+class ClientManager extends BaseClientManager
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function findClientByPublicId($publicId)
+    {
+        return $this->findClientBy(['randomId'  => $publicId]);
+    }
+}