+++ src/Controller/BackOffice/UserController.php
@@ -71,6 +71,9 @@ class UserController extends BackOfficeAbstractController
     /**
      * @Route("/manageUsers", name="manage_users")
      * @Security("is_granted('CAN_SHOW_USER')")
+
+     *
+     * @param Request $request
      */
     public function manageUsersAction(Request $request)
     {
@@ -127,6 +130,10 @@ class UserController extends BackOfficeAbstractController
             $this->getDoctrine()->getManager()->persist($user);
             $this->getDoctrine()->getManager()->flush();
 
+            if (isset($userData['sendResetingEmail'])) {
+                $this->sendResetingEmail($user);
+            }
+
             return $this->redirect($this->generateUrl('manage_users'));
         }
 
@@ -152,7 +159,6 @@ class UserController extends BackOfficeAbstractController
 
         if ($user->getProfile() === null) {
             $profile = new UserProfile();
-            $profile->setUser($user);
             $user->setProfile($profile);
         }
 
@@ -221,7 +227,8 @@ class UserController extends BackOfficeAbstractController
      * @Route("/manageUsers/sendResetPasswordEmail/{userId}", name="manage_users_send_reset_password_email")
      * @Security("is_granted('CAN_SHOW_USER')")
      *
-     * @param string $username
+     * @param Request $request
+     * @param string  $username
      */
     public function sendResetPasswordEmailAction(Request $request, $userId)
     {
@@ -245,12 +252,7 @@ class UserController extends BackOfficeAbstractController
             // Uncomment this « if / endif » to apply the TTL between to email sending
             // if (!$user->isPasswordRequestNonExpired($this->retryTtl)) {
 
-            if (null === $user->getConfirmationToken()) {
-                $user->setConfirmationToken($this->tokenGenerator->generateToken());
-            }
-            $this->mailer->sendResettingEmailMessage($user);
-            $user->setPasswordRequestedAt(new \DateTime());
-            $this->userManager->updateUser($user);
+            $this->sendResetingEmail($user);
 
             return $this->redirect($this->generateUrl('manage_users'));
 
@@ -261,6 +263,21 @@ class UserController extends BackOfficeAbstractController
         return $this->render('pia/User/sendResetPasswordEmail.html.twig', [
             'form' => $form->createView(),
         ]);
+    }
+
+    /**
+     * Sends FOSUser reset password email.
+     *
+     * @param UserInterface $user
+     */
+    private function sendResetingEmail(UserInterface $user): void
+    {
+        if (null === $user->getConfirmationToken()) {
+            $user->setConfirmationToken($this->tokenGenerator->generateToken());
+        }
+        $this->mailer->sendResettingEmailMessage($user);
+        $user->setPasswordRequestedAt(new \DateTime());
+        $this->userManager->updateUser($user);
     }
 
     protected function generateUsername(User $user)