+++ src/Controller/Pia/FolderController.php
@@ -24,12 +24,10 @@ class FolderController extends RestController
      * @FOSRest\Get("/folders")
      * @Security("is_granted('CAN_SHOW_FOLDER')")
      *
-     * @return array
+     * @return View
      */
     public function listAction(Request $request)
     {
-        $this->canAccessRouteOr304();
-
         $structure = $this->getUser()->getStructure();
         $collection = $this->getRepository()->findBy(['structure' => $structure, 'parent' => null]);
 
@@ -40,18 +38,16 @@ class FolderController extends RestController
      * @FOSRest\Get("/folders/{id}")
      * @Security("is_granted('CAN_SHOW_FOLDER')")
      *
-     * @return array
+     * @return View
      */
     public function showAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
         $folder = $this->getResource($id);
         if ($folder === null) {
             return $this->view($folder, Response::HTTP_NOT_FOUND);
         }
 
-        $this->canAccessResourceOr304($folder);
+        $this->canAccessResourceOr403($folder);
 
         return $this->view($folder, Response::HTTP_OK);
     }
@@ -60,12 +56,10 @@ class FolderController extends RestController
      * @FOSRest\Post("/folders")
      * @Security("is_granted('CAN_CREATE_FOLDER')")
      *
-     * @return array
+     * @return View
      */
     public function createAction(Request $request)
     {
-        $this->canAccessRouteOr304();
-
         if ($request->get('parent_id') === null && $request->get('parent') === null) {
             return $this->view('Missing parent identification', Response::HTTP_BAD_REQUEST);
         }
@@ -88,14 +82,12 @@ class FolderController extends RestController
      * @FOSRest\Post("/folders/{id}", requirements={"id"="\d+"})
      * @Security("is_granted('CAN_EDIT_FOLDER')")
      *
-     * @return array
+     * @return View
      */
     public function updateAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
         $folder = $this->getResource($id);
-        $this->canAccessResourceOr304($folder);
+        $this->canAccessResourceOr403($folder);
 
         $updatableAttributes = [
             'name'   => 'string',
@@ -113,14 +105,12 @@ class FolderController extends RestController
      * @FOSRest\Delete("/folders/{id}", requirements={"id"="\d+"})
      * @Security("is_granted('CAN_DELETE_FOLDER')")
      *
-     * @return array
+     * @return View
      */
     public function deleteAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
-        $folder = $this->getRepository()->find($id);
-        $this->canAccessResourceOr304($folder);
+        $folder = $this->getResource($id);
+        $this->canAccessResourceOr403($folder);
         $this->remove($folder);
 
         return $this->view($folder, Response::HTTP_OK);
@@ -131,7 +121,7 @@ class FolderController extends RestController
         return Folder::class;
     }
 
-    public function canAccessResourceOr304($resource): void
+    public function canAccessResourceOr403($resource): void
     {
         if (!$resource instanceof Folder) {
             throw new AccessDeniedHttpException();