+++ src/Controller/Pia/PiaController.php
@@ -43,8 +43,6 @@ class PiaController extends RestController
      */
     public function listAction(Request $request)
     {
-        $this->canAccessRouteOr304();
-
         $structure = $this->getUser()->getStructure();
 
         $criteria = array_merge($this->extractCriteria($request), ['structure' => $structure]);
@@ -61,14 +59,12 @@ class PiaController extends RestController
      */
     public function showAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
         $pia = $this->getRepository()->find($id);
         if ($pia === null) {
             return $this->view($pia, Response::HTTP_NOT_FOUND);
         }
 
-        $this->canAccessResourceOr304($pia);
+        $this->canAccessResourceOr403($pia);
 
         return $this->view($pia, Response::HTTP_OK);
     }
@@ -81,17 +77,15 @@ class PiaController extends RestController
      */
     public function createAction(Request $request)
     {
-        $this->canAccessRouteOr304();
-
         $pia = $this->newFromRequest($request);
 
-        if ($request->get('folder_id') !== null || $request->get('folder') !== null) {
-            $folderId = $request->get('folder_id', $request->get('folder')['id']);
-
+        if ($request->get('folder') !== null) {
+            $folderId = $request->get('folder')['id'];
             $folder = $this->getResource($folderId, Folder::class);
-            $pia->setFolder($folder);
+        } else {
+            $folder = $this->getUser()->getStructure()->getRootFolder();
         }
-
+        $pia->setFolder($folder);
         $pia->setStructure($this->getUser()->getStructure());
         $this->persist($pia);
 
@@ -106,7 +100,6 @@ class PiaController extends RestController
      */
     public function createFromTemplateAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
         /** @var PiaTemplate $piaTemplate */
         $piaTemplate = $this->getDoctrine()->getRepository(PiaTemplate::class)->find($id);
         if ($piaTemplate === null) {
@@ -114,6 +107,12 @@ class PiaController extends RestController
         }
 
         $pia = $this->jsonToEntityTransformer->transform($piaTemplate->getData());
+        if (($folderId = $request->get('folder')) !== null) {
+            $folder = $this->getResource($request->get('folder')['id'], Folder::class);
+        } else {
+            $folder = $this->getUser()->getStructure()->getRootFolder();
+        }
+        $pia->setFolder($folder);
         $pia->setName($request->get('name', $pia->getName()));
         $pia->setAuthorName($request->get('author_name', $pia->getAuthorName()));
         $pia->setEvaluatorName($request->get('evaluator_name', $pia->getEvaluatorName()));
@@ -133,17 +132,27 @@ class PiaController extends RestController
      */
     public function updateAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
         $pia = $this->getResource($id);
-        $this->canAccessResourceOr304($pia);
+        $this->canAccessResourceOr403($pia);
 
         $updatableAttributes = [
-            'name'           => 'string',
-            'author_name'    => 'string',
-            'evaluator_name' => 'string',
-            'validator_name' => 'string',
-        ];
+              'name'                              => 'string',
+              'author_name'                       => 'string',
+              'evaluator_name'                    => 'string',
+              'validator_name'                    => 'string',
+              'folder'                            => Folder::class,
+              'dpo_status'                        => 'int',
+              'concerned_people_status'           => 'int',
+              'status'                            => 'int',
+              'dpo_opinion'	                      => 'string',
+              'concerned_people_opinion'	         => 'string',
+              'concerned_people_searched_opinion' => 'boolean',
+              'concerned_people_searched_content' => 'string',
+              'rejection_reason'	                 => 'string',
+              'applied_adjustments'	              => 'string',
+              'dpos_names'                        => 'string',
+              'people_names'                      => 'sring',
+          ];
 
         $this->mergeFromRequest($pia, $updatableAttributes, $request);
 
@@ -154,16 +163,14 @@ class PiaController extends RestController
 
     /**
      * @FOSRest\Delete("/pias/{id}", requirements={"id"="\d+"})
-     * @Security("is_granted('CAN_EDIT_PIA')")
+     * @Security("is_granted('CAN_DELETE_PIA')")
      *
      * @return array
      */
     public function deleteAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
         $pia = $this->getRepository()->find($id);
-        $this->canAccessResourceOr304($pia);
+        $this->canAccessResourceOr403($pia);
         $this->remove($pia);
 
         return $this->view($pia, Response::HTTP_OK);
@@ -177,8 +184,6 @@ class PiaController extends RestController
      */
     public function importAction(Request $request)
     {
-        $this->canAccessRouteOr304();
-
         $importData = $request->get('data', null);
         if ($importData === null) {
             return $this->view($importData, Response::HTTP_BAD_REQUEST);
@@ -199,10 +204,10 @@ class PiaController extends RestController
      */
     public function exportAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
+        $this->canAccessRouteOr403();
 
         $pia = $this->getRepository()->find($id);
-        $this->canAccessResourceOr304($pia);
+        $this->canAccessResourceOr403($pia);
 
         $serializedPia = $this->jsonToEntityTransformer->reverseTransform($pia);
 
@@ -214,7 +219,7 @@ class PiaController extends RestController
         return Pia::class;
     }
 
-    public function canAccessResourceOr304($resource): void
+    public function canAccessResourceOr403($resource): void
     {
         if (!$resource instanceof Pia) {
             throw new AccessDeniedHttpException();