+++ src/Controller/Pia/PiaTemplateController.php
@@ -59,14 +59,12 @@ class PiaTemplateController extends RestController
      */
     public function showAction(Request $request, $id)
     {
-        $this->canAccessRouteOr304();
-
         $piaTemplate = $this->getRepository()->find($id);
         if ($piaTemplate === null) {
             return $this->view($piaTemplate, Response::HTTP_NOT_FOUND);
         }
 
-        $this->canAccessResourceOr304($piaTemplate);
+        $this->canAccessResourceOr403($piaTemplate);
 
         return $this->view($piaTemplate, Response::HTTP_OK);
     }
@@ -76,7 +74,7 @@ class PiaTemplateController extends RestController
         return PiaTemplate::class;
     }
 
-    public function canAccessResourceOr304($resource): void
+    public function canAccessResourceOr403($resource): void
     {
         if (!$resource instanceof PiaTemplate) {
             throw new AccessDeniedHttpException();