+++ src/Entity/Oauth/User.php
@@ -82,6 +82,7 @@ class User extends BaseUser implements AdvancedUserInterface, \Serializable
         $this->creationDate = new \DateTime();
         $this->expirationDate = new \DateTimeImmutable('+1 Year');
         $this->enabled = true;
+        $this->profile = new UserProfile();
     }
 
     /**
@@ -339,6 +340,7 @@ class User extends BaseUser implements AdvancedUserInterface, \Serializable
      */
     public function setProfile(?UserProfile $profile): void
     {
+        $profile->setUser($this);
         $this->profile = $profile;
     }
 }