+++ src/Form/Application/CreateApplicationForm.php
@@ -10,7 +10,7 @@
 
 namespace PiaApi\Form\Application;
 
-use Symfony\Component\Form\AbstractType;
+use PiaApi\Form\BaseForm;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\UrlType;
@@ -19,32 +19,32 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
 use OAuth2\OAuth2;
 
-class CreateApplicationForm extends AbstractType
+class CreateApplicationForm extends BaseForm
 {
     protected $grantTypes = [
-        OAuth2::GRANT_TYPE_AUTH_CODE          => OAuth2::GRANT_TYPE_AUTH_CODE,
-        OAuth2::GRANT_TYPE_IMPLICIT           => OAuth2::GRANT_TYPE_IMPLICIT,
-        OAuth2::GRANT_TYPE_USER_CREDENTIALS   => OAuth2::GRANT_TYPE_USER_CREDENTIALS,
-        OAuth2::GRANT_TYPE_CLIENT_CREDENTIALS => OAuth2::GRANT_TYPE_CLIENT_CREDENTIALS,
-        OAuth2::GRANT_TYPE_REFRESH_TOKEN      => OAuth2::GRANT_TYPE_REFRESH_TOKEN,
-        OAuth2::GRANT_TYPE_EXTENSIONS         => OAuth2::GRANT_TYPE_EXTENSIONS,
+        'pia.applications.labels.grant_types.' . OAuth2::GRANT_TYPE_AUTH_CODE          => OAuth2::GRANT_TYPE_AUTH_CODE,
+        'pia.applications.labels.grant_types.' . OAuth2::GRANT_TYPE_IMPLICIT           => OAuth2::GRANT_TYPE_IMPLICIT,
+        'pia.applications.labels.grant_types.' . OAuth2::GRANT_TYPE_USER_CREDENTIALS   => OAuth2::GRANT_TYPE_USER_CREDENTIALS,
+        'pia.applications.labels.grant_types.' . OAuth2::GRANT_TYPE_CLIENT_CREDENTIALS => OAuth2::GRANT_TYPE_CLIENT_CREDENTIALS,
+        'pia.applications.labels.grant_types.' . OAuth2::GRANT_TYPE_REFRESH_TOKEN      => OAuth2::GRANT_TYPE_REFRESH_TOKEN,
+        'pia.applications.labels.grant_types.' . OAuth2::GRANT_TYPE_EXTENSIONS         => OAuth2::GRANT_TYPE_EXTENSIONS,
     ];
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
             ->add('name', TextType::class, [
-                'label' => 'Nom',
+                'label' => 'pia.applications.forms.create.name',
             ])
             ->add('url', UrlType::class, [
-                'label' => 'URL',
+                'label' => 'pia.applications.forms.create.url',
             ])
             ->add('allowedGrantTypes', ChoiceType::class, [
                 'required' => false,
                 'multiple' => true,
                 'expanded' => true,
                 'choices'  => $this->grantTypes,
-                'label'    => 'Types d\'autorisations',
+                'label'    => 'pia.applications.forms.create.allowedGrantTypes',
             ])
             ->add('redirectUris', CollectionType::class, [
                 'entry_type'    => TextType::class,
@@ -55,13 +55,13 @@ class CreateApplicationForm extends AbstractType
                 'allow_delete'  => true,
                 'delete_empty'  => true,
                 'prototype'     => true,
-                'label'         => 'URIs de redirection',
+                'label'         => 'pia.applications.forms.create.redirectUris',
             ])
             ->add('submit', SubmitType::class, [
                 'attr' => [
                     'class' => 'fluid',
                 ],
-                'label' => 'Créer l\'application',
+                'label' => 'pia.applications.forms.create.submit',
             ])
         ;
     }