+++ tests/_support/WebGuy.php
@@ -12,14 +12,6 @@ use Facebook\WebDriver\WebDriverBy;
 use Facebook\WebDriver\Remote\RemoteWebElement;
 use Facebook\WebDriver\Remote\RemoteWebDriver;
 
-/*
- * Copyright (C) 2015-2018 Libre Informatique
- *
- * This file is licenced under the GNU LGPL v3.
- * For the full copyright and license information, please view the LICENSE.md
- * file that was distributed with this source code.
- */
-
 /**
  * Inherited Methods.
  *